Refactored project. Added TODO comments. In Routing.kt described work plan.
