+++ src/main/kotlin/team/microchad/Application.kt
@@ -3,9 +3,6 @@ package team.microchad
 import team.microchad.plugins.configureSerialization
 import io.ktor.server.application.*
 import io.ktor.server.config.*
-import kotlinx.coroutines.runBlocking
-import org.jetbrains.exposed.sql.Database
-import team.microchad.client.JiraClient
 import team.microchad.plugins.*
 
 fun main(args: Array<String>): Unit =
