+++ src/main/kotlin/team/microchad/config/MattermostConfiguration.kt
@@ -1,4 +1,7 @@
 package team.microchad.config
 
 class MattermostConfiguration {
+    //TODO get this values from application.yaml or PATH variables
+    val webhookUrl = "tin-workshop.ddns.net:8065/hooks"
+    val token = "5mj4ntowb7n3ddkb8hqbuw8sde"
 }
