--- src/main/kotlin/team/microchad/dto/ApiDto.kt
@@ -1,4 +1,0 @@
-package team.microchad.dto
-
-@kotlinx.serialization.Serializable
-data class SendToAll(val message: String)
