+++ src/main/kotlin/team/microchad/dto/jira/Field.kt
@@ -3,5 +3,5 @@ package team.microchad.dto.jira
 @kotlinx.serialization.Serializable
 data class Field(
     var summary: String,
-    var updated: String, //TODO исправить представление даты
+    var updated: String, //TODO( исправить представление даты)
 )
