+++ src/main/kotlin/team/microchad/dto/mm/IncomingMsg.kt
@@ -1,7 +1,7 @@
 package team.microchad.dto.mm
 
 import kotlinx.serialization.Serializable
-
+//TODO Rename this class
 @Serializable
 data class IncomingMsg(
     val channel_id: String,
