--- src/main/kotlin/team/microchad/dto/mm/User.kt
@@ -1,5 +1,0 @@
-package team.microchad.dto.mm
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class User(val id: String, val username: String)
