+++ src/main/kotlin/team/microchad/plugins/Koin.kt
@@ -1,10 +1,14 @@
 package team.microchad.plugins
 
 import io.ktor.server.application.*
+import org.koin.core.module.dsl.createdAtStart
+import org.koin.core.module.dsl.singleOf
+
 import org.koin.dsl.module
 import org.koin.ktor.plugin.Koin
 import org.koin.logger.slf4jLogger
 import team.microchad.client.MmClient
+import team.microchad.config.MattermostConfiguration
 
 fun Application.configureDI() {
     install(Koin) {
@@ -19,4 +23,7 @@ val mmBotModule = module {
         MmClient()
     }
 
+    singleOf(::MattermostConfiguration) {
+        createdAtStart()
+    }
 }
