+++ src/main/kotlin/team/microchad/plugins/Routing.kt
@@ -20,15 +20,12 @@ fun Application.configureRouting() {
         post("/") {
             var incomingMsg = call.receive<IncomingMsg>()
 
+            //TODO Описание работы
+            // parse jql from incoming msg using msgController. return Jql
+            // send jql using JiraClient. If all OK get JiraResponse and create outgoing msg
+            // else if some problems create msg with problem description
+            // send msg using mmClient
 
-
-
-//            val username = incomingMsg.user_name
-//            val status = MsgService(incomingMsg).getStatus()
-//            var jiraJqlResponse = JiraClient().getIssues(username, status)
-//            var outgoingMsg1 = OutgoingMsgConstructor(jiraJqlResponse, incomingMsg).getFormattedMessage()
-//            var outgoingMsg = OutgoingMsg(text = outgoingMsg1, channel = incomingMsg.channel_name, username = incomingMsg.user_name.plus("-").plus(status))
-//            MikeBot().send(outgoingMsg)
             MessageController().chooseCommand(incomingMsg)
 
         }
