--- src/main/kotlin/team/microchad/service/MsgService.kt
@@ -1,11 +1,0 @@
-package team.microchad.service
-
-import team.microchad.dto.mm.IncomingMsg
-
-class MsgService (message: IncomingMsg){
-    private var message = message;
-
-    fun getStatus(): String {
-        return message.text.replace("stats ", "")
-    }
-}
