+++ src/main/resources/application.yaml
@@ -9,4 +9,3 @@ bot:
     username: $BOT_USERNAME
     password: $BOT_PASSWORD
     mmtoken: $MM_TOKEN
-  key-word: stats
